+++ adapters/AdColony/AdColonyAdapter/GADMediationAdapterAdColony.m
@@ -24,12 +24,16 @@
 #import "GADMAdapterAdColonyInitializer.h"
 #import "GADMAdapterAdColonyRTBInterstitialRenderer.h"
 #import "GADMAdapterAdColonyRewardedRenderer.h"
+#import "GADMAdapterAdColonyRTBBannerRenderer.h"
 
 static AdColonyAppOptions *GADMAdapterAdColonyAppOptions;
 
 @implementation GADMediationAdapterAdColony {
   /// Completion handler for signal generation. Returns either signals or an error object.
   GADRTBSignalCompletionHandler _signalCompletionHandler;
+
+  /// AdColony banner ad renderer.
+  GADMAdapterAdColonyRTBBannerRenderer *_bannerRenderer;
 
   /// AdColony interstitial ad renderer.
   GADMAdapterAdColonyRTBInterstitialRenderer *_interstitialRenderer;
@@ -104,10 +108,6 @@ static AdColonyAppOptions *GADMAdapterAdColonyAppOptions;
   return [GADMAdapterAdColonyExtras class];
 }
 
-+ (GADVersionNumber)version {
-  return [GADMediationAdapterAdColony adapterVersion];
-}
-
 + (GADVersionNumber)adapterVersion {
   NSString *versionString = kGADMAdapterAdColonyVersionString;
   NSArray *versionComponents = [versionString componentsSeparatedByString:@"."];
@@ -124,22 +124,7 @@ static AdColonyAppOptions *GADMAdapterAdColonyAppOptions;
 
 - (void)collectSignalsForRequestParameters:(GADRTBRequestParameters *)params
                          completionHandler:(GADRTBSignalCompletionHandler)completionHandler {
-  // Keep handler, in practice this call may be asynchronous.
-  __block atomic_flag completionHandlerCalled = ATOMIC_FLAG_INIT;
-  __block GADRTBSignalCompletionHandler originalCompletionHandler = [completionHandler copy];
-  _signalCompletionHandler = ^void(NSString *_Nullable signals, NSError *_Nullable error) {
-    if (atomic_flag_test_and_set(&completionHandlerCalled)) {
-      return;
-    }
-
-    if (originalCompletionHandler) {
-      originalCompletionHandler(signals, error);
-    }
-    originalCompletionHandler = nil;
-  };
-    
-  NSString *signals = [AdColony collectSignals];
-  _signalCompletionHandler(signals, nil);
+  [AdColony collectSignals:completionHandler];
 }
 
 - (void)loadRewardedAdForAdConfiguration:(GADMediationRewardedAdConfiguration *)adConfiguration
@@ -158,5 +143,10 @@ static AdColonyAppOptions *GADMAdapterAdColonyAppOptions;
   [_interstitialRenderer renderInterstitialForAdConfig:adConfiguration
                                      completionHandler:completionHandler];
 }
-
+    
+- (void)loadBannerForAdConfiguration:(GADMediationBannerAdConfiguration *)adConfiguration
+completionHandler:(GADMediationBannerLoadCompletionHandler)completionHandler {
+    _bannerRenderer = [[GADMAdapterAdColonyRTBBannerRenderer alloc] init];
+    [_bannerRenderer renderBannerForAdConfig:adConfiguration completionHandler:completionHandler];
+}
 @end