+++ adapters/Chartboost/ChartboostAdapter/GADMediationAdapterChartboost.m
@@ -101,10 +101,6 @@
   return [GADMChartboostExtras class];
 }
 
-+ (GADVersionNumber)version {
-  return [GADMediationAdapterChartboost adapterVersion];
-}
-
 + (GADVersionNumber)adapterVersion {
   NSString *versionString = kGADMAdapterChartboostVersion;
   NSArray *versionComponents = [versionString componentsSeparatedByString:@"."];