+++ adapters/I-Mobile/IMobileAdapter/GADMediationAdapterIMobile.m
@@ -42,10 +42,6 @@
   return version;
 }
 
-+ (GADVersionNumber)version {
-  return [GADMediationAdapterIMobile adapterVersion];
-}
-
 + (GADVersionNumber)adapterVersion {
   GADVersionNumber version = {0};
   NSArray<NSString *> *components = [kGADMAdapterIMobileVersion componentsSeparatedByString:@"."];