+++ adapters/MoPub/MoPubAdapter/GADMAdapterMoPubConstants.h
@@ -3,7 +3,7 @@
 #define MAXIMUM_MOPUB_PRIVACY_ICON_SIZE 30
 
 /// MoPub mediation network adapter version.
-static NSString *const kGADMAdapterMoPubVersion = @"5.15.0.0";
+static NSString *const kGADMAdapterMoPubVersion = @"5.16.2.0";
 
 /// Constant for adapter error domain.
 static NSString *const kGADMAdapterMoPubErrorDomain = @"com.google.mediation.mopub";