+++ adapters/MoPub/MoPubAdapter/GADMediationAdapterMoPub.h
@@ -3,7 +3,7 @@
 
 typedef NS_ENUM(NSInteger, GADMoPubErrorCode) {
   /// The MoPub SDK sent a failure callback.
-  GADMoPubErrorSDKFailureCallback = 100,
+  // GADMoPubErrorSDKFailureCallback = 100,
   /// An ad is already loaded for this network configuration.
   GADMoPubErrorAdAlreadyLoaded = 101,
   /// There was an error loading data from the network.