+++ adapters/MyTarget/MyTargetAdapter/GADMAdapterMyTargetConstants.m
@@ -14,7 +14,7 @@
 
 #import "GADMAdapterMyTargetConstants.h"
 
-NSString *const _Nonnull kGADMAdapterMyTargetVersion = @"5.9.9.0";
+NSString *const _Nonnull kGADMAdapterMyTargetVersion = @"5.10.1.0";
 
 NSString *const _Nonnull kGADMAdapterMyTargetSlotIdKey = @"slotId";