+++ adapters/Fyber/CHANGELOG.md
@@ -1,5 +1,30 @@
 ## Fyber iOS Mediation Adapter Changelog
 
+#### Version 7.8.2.0
+- Verified compatibility with Fyber Marketplace SDK version 7.8.2.
+- Now requires Google Mobile Ads SDK version 8.2.0 or higher.
+
+Built and tested with:
+- Google Mobile Ads SDK version 8.2.0.
+- Fyber Marketplace SDK version 7.8.2.
+
+#### Version 7.8.1.0
+- Verified compatibility with Fyber Marketplace SDK version 7.8.1.
+- Now requires Google Mobile Ads SDK version 8.0.0 or higher.
+- Added standardized adapter error codes and messages.
+- Updated the adapter to use the `.xcframework` format.
+
+Built and tested with:
+- Google Mobile Ads SDK version 8.0.0.
+- Fyber Marketplace SDK version 7.8.1.
+
+#### Version 7.8.0.0
+- Verified compatibility with Fyber Marketplace SDK version 7.8.0.
+
+Built and tested with:
+- Google Mobile Ads SDK version 7.69.0.
+- Fyber Marketplace SDK version 7.8.0.
+
 #### Version 7.7.3.0
 - Verified compatibility with Fyber Marketplace SDK version 7.7.3.
 - Now requires Google Mobile Ads SDK version 7.69.0 or higher.