+++ adapters/Unity/CHANGELOG.md
@@ -7,6 +7,15 @@ Built and tested with
 - Google Mobile Ads SDK version 7.69.0.
 - Unity Ads SDK version 3.7.1.
 
+#### Version 3.6.2.0
+- Updated the adapter to use the `.xcframework` format.
+- Verified compatibility with unity ads SDK 3.6.2.
+- Now requires Google Mobile Ads SDK version 8.2.0 or higher.
+
+Built and tested with
+- Google Mobile Ads SDK version 8.2.0.
+- Unity Ads SDK version 3.6.2.
+
 #### Version 3.6.0.0
 - Verified compatibility with unity ads SDK 3.6.0.
 - Now requires Google Mobile Ads SDK version 7.69.0 or higher.